+++ R/elicitConcProb.R
@@ -1,4 +1,4 @@
-#' Elicit a quadrant probability for two uncertain quantities, and plot a joint sample
+#' Elicit a concordance probability for two uncertain quantities, and plot a joint sample
 #' 
 #' Given two elicited marginal distributions, open a browser in which one specifies a quadrant
 #' probability P(X_1 > m_1, X_2 > m_2), where m_1 and m_2 are the elicited medians of X_1 and X_2.
@@ -28,29 +28,41 @@
 #' elicitQuadProb(myfit1, myfit2, 0.55, 0.3, d=c("Beta", "Beta"))
 #' }
 #' @import shiny
+#' @importFrom ggExtra ggMarginal
 #' @export
 
-elicitQuadProb <- function(fit1, fit2, m1, m2, 
+elicitConcProb <- function(fit1, fit2, m1, m2, 
                               d = c("best", "best"), 
                               n = 10000){
-  qplabel <- paste("\\( P(X_1>", m1, ",X_2 >", m2,")\\)")
+  qplabel <- paste("\\( P(X_1>", m1,
+                   ",X_2 >", m2,
+                   " \\mbox{ or } X_1 <", m1, 
+                   ",X_2 <", m2,
+                   ")\\)")
+  conc.probs <- matrix(0.5, 2, 2)
+  theta<-data.frame(copulaSample(fit1, fit2, cp=conc.probs, n=n, 
+                                 d=d))
+  xrange <- range(theta$X1)
+  yrange <- range(theta$X2)
+  
   ui <- basicPage(
     withMathJax(), 
-    titlePanel("Elicit a quadrant probability"),
-    sliderInput("topright", label = h4(qplabel), 
-                0.25, min = 0, max = 0.5, step = 0.01),
+    titlePanel("Elicit a concordance probability"),
+    sliderInput("cprob", label = h4(qplabel), 
+                0.5, min = 0, max = 1, step = 0.01),
     plotOutput("plot1"),
     numericInput("fs", label = h4("font size"), value = 12)
   )
   
   server <- function(input, output) {
     
+    X1 <- X2 <- xpos <- ypos <- hjustvar <- vjustvar <- annotateText <- NULL # hack to avoid R CMD check NOTE
     
     output$plot1 <- renderPlot({
-      quad.probs <- matrix(0, 2, 2)
-      quad.probs[1, 2] <- input$topright
+      conc.probs <- matrix(0, 2, 2)
+      conc.probs[1, 2] <- input$cprob
       set.seed(1)
-      theta<-data.frame(copulaSample(myfit1, myfit2, qp=quad.probs, n=n, 
+      theta<-data.frame(copulaSample(fit1, fit2, cp=conc.probs, n=n, 
                                      d=d))
     
       df1<-data.frame(theta)
@@ -59,14 +71,15 @@ elicitQuadProb <- function(fit1, fit2, m1, m2,
       annotations <- data.frame(
         xpos = c(Inf,Inf,-Inf,-Inf),
         ypos =  c(Inf, -Inf,-Inf,Inf),
-        annotateText = as.character(c(input$topright, 0.5 - input$topright,
-                                      input$topright, 
-                                      0.5 - input$topright)),
+        annotateText = as.character(c(input$cprob / 2, 0.5 - input$cprob /2,
+                                      input$cprob / 2, 
+                                      0.5 - input$cprob /2)),
         hjustvar = c(1.5, 1.5, -0.5, -0.5) ,
         vjustvar = c(1.5, -0.5, -0.5, 1.5))
       
       
-      ggplot(data=df1,aes(x=X1, y=X2))+geom_point(alpha=0.15, colour = "red") +
+      p1<-ggplot(data=df1,aes(x=X1, y=X2))+
+        geom_point(alpha=0.15, colour = "red") +
         geom_hline(yintercept = m2)+
         geom_vline(xintercept = m1)+
         labs(x=expression(X[1]), y = expression(X[2]))+
@@ -75,8 +88,11 @@ elicitQuadProb <- function(fit1, fit2, m1, m2,
                                           hjust = hjustvar,
                                           vjust = vjustvar,
                                           label = annotateText),
-                  size =10)
-       
+                  size =10)+
+        xlim(0.95*xrange[1], 1.05*xrange[2])+
+        ylim(0.95*yrange[1], 1.05*yrange[2])
+      suppressWarnings(suppressMessages(ggExtra::ggMarginal(p1, type = "histogram",
+                                           fill = "red")))
     })
   }